+++ lib/arctium.dart
@@ -1,16 +1,12 @@
 library arctium;
 
-import 'dart:async';
-import 'package:arctium/src/private/duplicatable_map.dart';
-import 'package:flutter/material.dart';
-
 // Private
-part 'src/private/_dispose_queue.dart';
+export 'src/private/dispose_queue.dart';
 
 // Public
-part 'src/public/disposable.dart';
-part 'src/public/arctium.dart';
+export 'src/public/disposable.dart';
+export 'src/public/arctium.dart';
 
 // Widgets
-part 'src/widgets/locator_widget.dart';
-part 'src/widgets/injector.dart';
+export 'src/widgets/locator_widget.dart';
+export 'src/widgets/injector.dart';