+++ lib/src/widgets/locator_widget.dart
@@ -1,4 +1,5 @@
-part of arctium;
+import 'package:flutter/material.dart';
+import 'package:arctium/src/public/arctium.dart';
 
 /// Widget that locates service under the given type parameter.
 ///
@@ -6,7 +7,7 @@ part of arctium;
 /// it as a parameter in the [build] method. Locating it by hand before
 /// returning from build is equivalent, but requires more code.
 abstract class LocatorWidget<T> extends StatefulWidget {
-  const LocatorWidget();
+  const LocatorWidget({Key? key}) : super(key: key);
 
   Widget build(BuildContext context, T service);