+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.2.0] - 09.08.2021
+
+- Added keys to Widgets
+- [BREAKING CHANGES] Changed DisposeQueue behavior from sync to async
+- [BREAKING CHANGES] Changed Injector service parameter to lazy version
+
 ## [0.1.2] - 29.06.2021
 
 - Added const constructor to LocatorWidget